% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{birth}
\alias{birth}
\title{Birth pulse functions}
\usage{
birth(t, s, omega, k)
}
\arguments{
\item{t}{time in years}

\item{s}{synchrony}

\item{omega}{offset}

\item{k}{scale of function}
}
\value{
vector
}
\description{
Birth pulse functions
}
