% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{conec.dist}
\alias{conec.dist}
\title{Connectivity function}
\usage{
conec.dist(dist, d1 = 5e-04, d2 = 10000)
}
\arguments{
\item{dist}{pairwise distance}

\item{d1}{sigmoid slope}

\item{d2}{distance threshold}
}
\value{
vector
}
\description{
Connectivity function
}
