% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dist_col_URY}
\alias{dist_col_URY}
\title{Connetivity matrix}
\format{
A matrix with 25 rows and 25 columns:
}
\usage{
dist_col_URY
}
\description{
A dataset containing the initial connectivity matrix for rmpmr_seir function
based on uruguayan dataset of colonies (Botto 2021)
}
\seealso{
Other data: 
\code{\link{pop}}
}
\concept{data}
\keyword{datasets}
