% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_par.R
\name{initial_par}
\alias{initial_par}
\title{initial_par}
\usage{
initial_par(pop, dist, init_infect = 5, init_cols = 1, itera = 1, ...)
}
\arguments{
\item{pop}{matrix with 7 columns corresponding to suscepts, exp, infect, recov, time, newborns, deaths}

\item{dist}{squear matrix of conectivity among colonies}

\item{init_infect}{number of initial infect. Default 5}

\item{init_cols}{number of initial colonies Default 1}

\item{itera}{number of iterations. Default 1}

\item{...}{All other param of rmpmr_seir}
}
\value{
list of lists
}
\description{
initial_par
}
\examples{
\dontrun{
 modelados <- initial_par(pop, dist_col_URY)
}
}
