% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.R
\name{persistence}
\alias{persistence}
\title{persistence}
\usage{
persistence(pop, dist, threshold = seq(0, 1e+06, 1e+05), iter = 11, ...)
}
\arguments{
\item{pop}{matrix with 7 columns corresponding to suscepts, exp, infect, recov, time, newborns, deaths}

\item{dist}{squear matrix of conectivity among colonies}

\item{threshold}{each distance threshold to eval. Default seq(0, 1000000, 100000)}

\item{iter}{number of iterations. Default 11}

\item{...}{All other param of rmpmr_seir}
}
\value{
list of lists
}
\description{
persistence
}
\examples{
\donttest{
 persistence(pop, dist_col_URY)
}
}
