% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop}
\alias{pop}
\title{Population matrix}
\format{
A matrix with 25 rows and 7 columns:
\describe{
\item{suscept}{number of population susceptible at time 0}
\item{exp}{number of population expoussed at time 0}
\item{infect}{number of population infected at time 0}
\item{recov}{number of population recovered at time 0}
\item{time}{inicial time)}
\item{newborns}{number of population born at time 0}
\item{deaths}{number of population death at time 0}
}
}
\usage{
pop
}
\description{
A dataset containing the initialpopulation matrix for rmpmr_seir function
}
\seealso{
Other data: 
\code{\link{dist_col_URY}}
}
\concept{data}
\keyword{datasets}
