% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{r_t}
\alias{r_t}
\title{Reproductivity success over time}
\usage{
r_t(t, s = 0.4, omega = 4.55, glob.rep = 0.102/365.25)
}
\arguments{
\item{t}{time in days}

\item{s}{synchrony}

\item{omega}{offset}

\item{glob.rep}{total reproductive rate}
}
\value{
vector
}
\description{
Reproductivity success over time
}
