% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-package.R
\docType{package}
\name{rmpmr-package}
\alias{rmpmr}
\alias{rmpmr-package}
\title{rmpmr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Richard Detomasi \email{richard.detomasi@gmail.com}

Authors:
\itemize{
  \item Germán Botto \email{german.botto@gmail.com}
}

}
\keyword{internal}
