% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpmr_seir.R
\name{rmpmr_seir}
\alias{rmpmr_seir}
\title{Dinamic model}
\usage{
rmpmr_seir(
  pop,
  dist,
  dt = 1,
  ntt = 18250,
  glob.rep = 0.102/365.25,
  mu = glob.rep,
  omega = 4.55,
  s = 0.4,
  epsilon = 1/135,
  delta = 0.1,
  tau = 1/21,
  alpha = 1/11,
  beta = 0.9322064,
  dens = T
)
}
\arguments{
\item{pop}{matrix with 7 columns corresponding to suscepts, exp, infect, recov, time, newborns, deaths}

\item{dist}{squear matrix of conectivity among colonies}

\item{dt}{Time step in days. Default 1}

\item{ntt}{Max. limit of times in days. Default 18250 (aprox. 50 years)}

\item{glob.rep}{Total reproductive rate (1/days). Default  0.102/365.25}

\item{mu}{Natural mortality rate (1/days). Default glob.rep}

\item{omega}{Offset. Default 4.55}

\item{s}{Synchrony. Default 0.4}

\item{epsilon}{Immunity waining rate (1/days). Default 1/135}

\item{delta}{Proportion of exposued to devlop infection. Default 0.1}

\item{tau}{Exit rate from exposed class. Default 1/21}

\item{alpha}{Disease-induced mortality rate. Default 1/11}

\item{beta}{probability of transmission given contact. Default 0.9322064}

\item{dens}{If T by density dependent transmission, else by frequency dependent transmission. Default T}
}
\value{
list of lists
}
\description{
Dinamic model
}
\examples{
\dontrun{
 rmpmr_seir(pop, dist_col_URY)
}
}
